requirements.txt
# Core Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1

# Database
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1

# AI/ML - Face Recognition
# Updated for Python 3.10+ compatibility
tensorflow>=2.16.1
tf-keras>=2.16.0
deepface>=0.0.90
mtcnn>=0.1.1

# Computer Vision
opencv-python>=4.9.0.80
opencv-contrib-python>=4.9.0.80
Pillow>=10.1.0

# Face Recognition Libraries
face-recognition>=1.3.0
dlib>=19.24.2; platform_system != "Windows"

# Numerical Computing
numpy>=1.26.2
scipy>=1.11.4

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Video Processing
moviepy==1.0.3
imageio==2.33.1
imageio-ffmpeg==0.4.9

# Progress Tracking
tqdm==4.66.1

# Data Handling
pandas==2.1.4

# API & Validation
marshmallow==3.20.1
python-multipart==0.0.6
